flows: respect authentication_classes in FlowExecutorView()
With this commit, we move the session authentication logic from
the dispatch() method to perform_authentication() method, which
is now also overridden from the APIView parent class.

The major difference here is the order in which things are called
now.

The most important point is that the django request object is
wrapped by the django rest framework in APIView.dispatch() by
calling APIView.initialize_request(). This method returns a
wrapped version of the request (of type rest_framework.Request).
This wrapper class implicitly performs the authentication
from the authentication_classes, when request.user is accessed.

Before this commit, the original django request object
was passed to FlowPlanner._check_authentication() before
it was wrapped into a rest_framework.Request, which led
to the authentication_classes not been recognized.

The call chain was:
    FlowExecutorView.dispatch()
      FlowExecutorView._initiate_plan()
        FlowPlanner.plan()
          FlowPlanner._check_authentication()
      APIView.dispatch()
    APIView.initialize_request()

With this commit, the flow session logic is moved to
FlowExecutorView.perform_authentication() such that
the wrapped request instance is properly passed down
to FlowPlanner._check_authentication()

The new call chain is then:
    FlowExecutorView.dispatch()
      APIView.dispatch()
        APIView.initialize_request()
        APIView.initial()
          FlowExecutorView.perform_authentication()
            FlowExecutorView._initiate_plan()
              FlowPlanner.plan()
                FlowPlanner._check_authentication()

Since we can not return HttpResponses from
FlowExecutorView.perform_authentication() some exception
handlers remain in the dispatch() method. Additionally,
two new exceptions are introduced which are raised
in perform_authentication() and caught in dispatch() to
be transformed into HttpResponses.

Fixes #15747